requirements.txt
# FastAPI and server
fastapi>=0.128.0
uvicorn[standard]
pydantic>=2.12.0
pydantic-settings>=2.12.0
httpx>=0.28.1

# YouTube integration
youtube-transcript-api>=0.128.0
yt-dlp>=2025.12.8
faster-whisper>=1.1.1

# MCP (Model Context Protocol) server
fastmcp>=2.14.1

# Rate limiting
slowapi

# Testing
pytest
pytest-asyncio
httpx

# Development
python-dotenv
